requirements.txt
alembic==1.16.2
altgraph==0.17.4
arabic-reshaper==3.0.0
asn1crypto==1.5.1
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
dotenv==0.9.9
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fpdf==1.7.2
greenlet==3.2.3
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.11.0.86
oscrypto==1.3.0
packaging==25.0
pdf2image==1.17.0
pefile==2023.2.7
pillow==11.2.1
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.10.1
pypdf==5.7.0
python-bidi==0.6.6
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
setuptools==80.8.0
six==1.17.0
SQLAlchemy==2.0.41
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
xhtml2pdf==0.2.17
